refactor: move routers to their own feature
