+++ src/router/index.ts
@@ -1,8 +1,8 @@
 import Router, { RouterContext } from '@koa/router';
 import jwt from 'koa-jwt';
 import { environment } from '@/config/environment';
-import { usersRouter } from './users.router';
-import { authRouter } from './auth.router';
+import { usersRouter } from '../features/users/users.router';
+import { authRouter } from '../features/auth/auth.router';
 
 export const router = new Router();
